Change battery full level from 90 to 100
*DU Notes*

Originally this was set to mirror the existing BATTERY_LIGHT_FULL_COLOR mod
by changing the battery level from 90 to 100 and in turn having 4 selections
in Settings.

Batery Low
Charging
Charged (90%)
Charged (100%)

You can see here under the first patchset
http://gerrit.dirtyunicorns.com/#/c/29253/1/

This made no sense to me once I really took a look at the code in here. It seem
unnecessary. If we're modifiying this to be a mod where we can set the charge level
or maybe toggle between 90 and 100 then OK but as it stands, this was basically
just changing the level so this is what I changed it to.

Change-Id: I12de871559d1cfbc4eeba3dbe7f5aec24b2338cc